fix: rename run-dev-local to run-local-runtime
